Rework email verification component
