+++ src/components/features/auth/EmailVerificationCallout.tsx
@@ -1,0 +1,42 @@
+"use client";
+
+import { Callout } from "@radix-ui/themes";
+import { CheckCircledIcon, InfoCircledIcon } from "@radix-ui/react-icons";
+import { useSearchParams } from "next/navigation";
+
+interface EmailVerificationCalloutProps {
+  showCheckEmail?: boolean;
+}
+
+export function EmailVerificationCallout({
+  showCheckEmail = false,
+}: EmailVerificationCalloutProps) {
+  const searchParams = useSearchParams();
+  const isVerified = searchParams.has("verified");
+
+  if (!isVerified || showCheckEmail) {
+    return (
+      <Callout.Root color="blue">
+        <Callout.Icon>
+          <InfoCircledIcon />
+        </Callout.Icon>
+        <Callout.Text>
+          Please check your email. We&apos;ve sent you a code to verify your
+          email address.
+        </Callout.Text>
+      </Callout.Root>
+    );
+  }
+
+  return (
+    <Callout.Root color="green">
+      <Callout.Icon>
+        <CheckCircledIcon />
+      </Callout.Icon>
+      <Callout.Text>
+        Your email has been successfully verified. Thank you for confirming your
+        account.
+      </Callout.Text>
+    </Callout.Root>
+  );
+}
